requirements_test.txt
# Warning: For automatic run of test in Gitlab CI, we must not include other things that pytest-homeassistant-custom-component
# full reload dependencies: pip cache purge && pip install --use-feature=fast-deps -r requirements_test.txt
# list all version of pytest-homeassistant-custom-component: pip index versions pytest-homeassistant-custom-component
-r requirements_dev.txt
aiodiscover
ulid_transform
pytest
coverage
pytest-asyncio
pytest-homeassistant-custom-component
pytest-socket
# pycares >=4.4.0 required for aiodns typing support on Python 3.13 (missing ares_query_a_result otherwise)
pycares>=4.4.0
